requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
backports.zoneinfo==0.2.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==4.0
django-celery-beat==2.7.0
django-timezone-field==7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
kombu==5.4.2
prompt-toolkit==3.0.50
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.1
redis==5.2.1
redis-server==6.0.9
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.12.2
tzdata==2025.1
vine==5.1.0
virtualenv==20.29.2
wcwidth==0.2.13
whitenoise==6.7.0
django-cors-headers
